requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==5.8.0
idna==2.10
inflection==0.5.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.8
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pudb==2020.1
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.3
urwid==2.1.2
wcwidth==0.2.5
